pyproject.toml
[project]
name = "sfmc-stg-migrate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[clickhouse,postgres]>=1.21.0",
    "google>=3.0.0",
    "google-cloud-bigquery>=3.40.0",
    "pandas>=3.0.0",
    "pandas-gbq>=0.31.0",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]
